Implement job posting command
This commit implements the job posting command for user of the server to
post jobs with the link, company, title, experience level, and job
description. This command is set up to be used in one command, specified
by name in the job_channel field in the config. The command will
respond in the job_board field in teh config, specified by the channel
ID. The usage is !job url | company | title | experience level |
description.